Creating class and test

+++ src/main/kotlin/digionebank/Conta.kt
@@ -1,16 +1,14 @@
 package digionebank
 
-import java.math.BigDecimal
-
 class Conta (
     val agencia: String,
     val numero: String,
-    val saldo: BigDecimal
+    val saldo: Double
 ){
-    fun deposito(valor: BigDecimal){
+    fun deposito(valor: Double){
 
     }
-    fun saque(valor: BigDecimal){
+    fun saque(valor: Double){
 
     }
 }
